factored widget out of App
